agents/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ailine-agents"
version = "0.1.0"
description = "AiLine agent framework: Pydantic AI agents + LangGraph workflows"
requires-python = ">=3.11"
dependencies = [
  "pydantic-ai>=1.58.0,<2",
  "langgraph==1.0.8",
  "pydantic>=2.10,<3",
  "pydantic-settings>=2.7,<3",
  "structlog>=24,<26",
  "deepagents==0.4.1",
  "langchain-anthropic>=1.3.2,<2",
  "langchain-core>=1.2.10,<2",
  "anthropic>=0.79.0,<1",
  "fastapi>=0.128,<1",
  "sse-starlette>=3.2,<4",
  "httpx>=0.28,<1",
  "tenacity>=9,<10",
  "numpy>=1.26,<3",
  "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
anthropic = ["pydantic-ai[anthropic]"]
openai = ["pydantic-ai[openai]"]
gemini = ["pydantic-ai[google]"]
all = ["ailine-agents[anthropic,openai,gemini]"]

[dependency-groups]
dev = [
  "pytest>=8,<9",
  "pytest-asyncio>=0.24,<1",
  "pytest-cov>=6,<7",
  "ruff>=0.9,<1",
  "mypy>=1.14,<2",
  "aiosqlite>=0.20,<1",
]

[tool.hatch.build.targets.wheel]
packages = ["ailine_agents"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
  "live_llm: requires real LLM API keys",
]

[tool.coverage.run]
source = ["ailine_agents"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if __name__",
  "if TYPE_CHECKING",
]
